Fix warning when building with newlib
If we use newlib the isdigit (and other similar functions) return an
error as char can possibly be viewed as signed:

usr/include/ctype.h:57:54: error: array subscript has type ‘char’ [-Werror=char-subscripts]
 #define __ctype_lookup(__c) ((__ctype_ptr__+sizeof(""[__c]))[(int)(__c)])

Explicity cast to unsigned char so we deal with both this warning and
possible warning when -Wpointer-sign is enabled.

+++ source/src/Json/codec/JsonTokenizer.c
@@ -248,7 +248,7 @@ static TokenParseResult JsonTokenizer_ParseString(JsonTokenizer *thiz)
                 break;
             }
 
-            else if (isprint(c))
+            else if (isprint((int)c))
             {
                 thiz->current ++;
             }
@@ -315,7 +315,7 @@ static TokenParseResult JsonTokenizer_ParseNumber(JsonTokenizer *thiz)
                 thiz->current ++;
                 c = *thiz->current;
 
-                while (isdigit(c))
+                while (isdigit((int)c))
                 {
                     thiz->current ++;
                     c = *thiz->current;
@@ -342,7 +342,7 @@ static TokenParseResult JsonTokenizer_ParseNumber(JsonTokenizer *thiz)
                 break;
             }
 
-            while (isdigit(c))
+            while (isdigit((int)c))
             {
                 thiz->current ++;
                 c = *thiz->current;
@@ -359,7 +359,7 @@ static TokenParseResult JsonTokenizer_ParseNumber(JsonTokenizer *thiz)
 
                 if (c == '+' || c == '-')
                 {
-                    while (isdigit(c))
+                    while (isdigit((int)c))
                     {
                         thiz->current++;
                         c = *thiz->current;
